requirements.txt
bcrypt==3.2.0
click==8.1.3
starlette~=0.20.0
PyJWT~=2.4.0
uvicorn~=0.17.6